debug-assignment/blood-test-analyser-debug/requirements.txt
# Core dependencies - Install CrewAI from GitHub
git+https://github.com/joaomdmoura/crewAI.git
crewai-tools==0.47.1
fastapi==0.110.3
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1

# LangChain and LLM providers
langchain-core==0.1.52
langchain-openai==0.1.25
langchain-google-genai==0.1.8
langchain-community==0.0.27
langsmith==0.1.67

# OpenAI and Google AI
openai==1.30.5
google-generativeai==0.5.4

# Data processing
pandas==2.2.2
numpy==1.26.4

# Utilities
click>=8.0.0
Jinja2>=3.0.0
jsonschema>=4.0.0
protobuf>=4.0.0
pydantic>=1.10.0
pydantic_core>=2.8.0

# Optional: Google Cloud (if using Google AI)
google-api-core>=2.0.0
google-auth>=2.0.0
google-cloud-aiplatform>=1.50.0